Refactor du code pour replacer les classes pour séparer les fonctionalités de chaque couche de domain.
