+++ DeskBooker.Core.Tests/DeskBooker.Core.Tests.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="coverlet.collector" Version="1.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DeskBooker.Core\DeskBooker.Core.csproj" />
+  </ItemGroup>
+
 </Project>
