+++ DeskBooker.Core.Tests/Processor/DeskBookingRquestProcessorTests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using DeskBooker.Core.Domain;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using Xunit;
