+++ DeskBooker.Core/Domain/DeskBookingResult.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace DeskBooker.Core.Processor
+
+namespace DeskBooker.Core.Domain
 {
-    internal class DeskBookingRequest
+    public class DeskBookingResult
     {
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Email { get; set; }
         public DateTime Date { get; set; }
-
 
     }
 }
