+++ DeskBooker.Core/Processor/DeskBookingRequestProcessor.cs
@@ -1,14 +1,15 @@
-﻿using System;
+﻿using DeskBooker.Core.Domain;
+using System;
 
 namespace DeskBooker.Core.Processor
 {
-    internal class DeskBookingRequestProcessor
+    public class DeskBookingRequestProcessor
     {
         public DeskBookingRequestProcessor()
         {
         }
 
-        internal DeskBookingResult BookDesk(DeskBookingRequest request)
+        public DeskBookingResult BookDesk(DeskBookingRequest request)
         {
             return new DeskBookingResult
             {
